+++ R/eng-mod_ts_prv.R
@@ -25,7 +25,7 @@ eng_mod_ts_prv_ui <- function(id){
         )
       )
     ),
-    fluidRow(shiny::checkboxInput(ns("y_log"), "Scala logaritmica")),
+    fluidRow(shiny::checkboxInput(ns("y_log"), "Logarithmic scale")),
     fluidRow(plotlyOutput(ns("ts_plot"), height = "200%"))
   )
 }
@@ -96,7 +96,9 @@ eng_mod_ts_prv_server <- function(id, type = c("cum", "inc")) {
           ylab(paste0(y_lab()," - log2"))
       }
 
-      ggplotly(gg)
+      ggplotly(gg) %>%
+        config(modeBarButtonsToRemove = c("zoomIn2d", "zoomOut2d", "pan2d", "select2d", "lasso2d")) %>%
+        config(displaylogo = FALSE)
     })
 
   })